Documentation and API tweaks
