+++ FlintCore/Constraints/Permissions/SystemPermissionRequestAction.swift
@@ -8,8 +8,19 @@
 
 import Foundation
 
+/// The type used to describe what to do with the current permission request during
+/// an authorisation controller flow. Your `PermissionAuthorisationCoordinator` passes these values
+/// to the completion handler of `willRequestPermission(:completion:)` to indicate what the controller
+/// should do next.
+///
+/// - see: `PermissionAuthorisationCoordinator`
 public enum SystemPermissionRequestAction {
-    case requestPermission
-    case skipPermission
+    /// Continue and show the system permission request alert
+    case request
+    
+    /// Skip this permission - perhaps because the user tapped "Not now" in your custom onboarding UI
+    case skip
+    
+    /// Cancel the entire authorisation flow. `didCompletePermissionAuthorisation` will be called ont he coordinator.
     case cancelAll
 }
